r/aws_kinesisanalyticsv2_application_snapshot: Reduce visibility.
