--- internal/service/kinesisanalyticsv2/id.go
@@ -1,28 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package kinesisanalyticsv2
-
-import (
-	"fmt"
-	"strings"
-)
-
-const applicationSnapshotIDSeparator = "/"
-
-func applicationSnapshotCreateID(applicationName, snapshotName string) string {
-	parts := []string{applicationName, snapshotName}
-	id := strings.Join(parts, applicationSnapshotIDSeparator)
-
-	return id
-}
-
-func applicationSnapshotParseID(id string) (string, string, error) {
-	parts := strings.Split(id, applicationSnapshotIDSeparator)
-
-	if len(parts) == 2 && parts[0] != "" && parts[1] != "" {
-		return parts[0], parts[1], nil
-	}
-
-	return "", "", fmt.Errorf("unexpected format for ID (%q), expected application-name%ssnapshot-name", id, applicationSnapshotIDSeparator)
-}
